Clean up codes

+++ test1.cpp
@@ -7,9 +7,8 @@ int main() {
     cin >> tt;
     Quiz1 q1;
     while (tt--) {
-        long long w, x, y, z;
+        int w, x, y, z;
         cin >> w >> x >> y >> z;
-
         cout << q1.program(w, x, y, z) << '\n';
     }
